Streamline option parsing code
This commit tries to untangle the previous complex option parsing code
and simplify flow control understanding.

Besides, the `-m` option does not accept only the first letter of the
desired mode (c, i or t) anymore, as it was not documented and it
presented a potential for errors as users mistyped the mode.